requirements_web.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.18
aiofiles>=24.0
requests
opencv-python
ffmpeg-python
Pillow
